Move class to avoid cyclic package dependency
